+++ src/schema/article-renderer.tsx
@@ -16,13 +16,13 @@ import { Solution } from '@/components/content/exercises/solution'
 import { Geogebra } from '@/components/content/geogebra'
 import { Injection } from '@/components/content/injection'
 import { Lazy } from '@/components/content/lazy'
-import { MathSpanProps } from '@/components/content/math-span'
+import type { MathSpanProps } from '@/components/content/math-span'
 import { Multimedia } from '@/components/content/multimedia'
 import { SerloTable } from '@/components/content/serlo-table'
 import { Snack } from '@/components/content/snack'
 import { Spoiler } from '@/components/content/spoiler'
 import { Video } from '@/components/content/video'
-import { FrontendContentNode } from '@/data-types'
+import type { FrontendContentNode } from '@/data-types'
 
 export type NodePath = (number | string)[]
